Refactor for better testability (factory pattern for app as fixture)
